functions/schemas_api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via google-crc32c
chardet==4.0.0
    # via requests
google-api-core==1.30.0
    # via google-cloud-core
google-auth==1.32.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==1.7.1
    # via google-cloud-storage
google-cloud-storage==1.40.0
    # via -r requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.3.1
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
idna==2.10
    # via requests
packaging==20.9
    # via google-api-core
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via google-api-core
requests==2.25.1
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   google-api-core
    #   google-auth
    #   google-cloud-core
    #   google-resumable-media
    #   protobuf
urllib3==1.26.6
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
